samples/ingestion/ingestion-client/OpenAIFunction/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy
requests
pandas
azure-storage-blob 
azure-identity
smart_open
openai
tenacity
pinecone-client
redis